Rename module to biomechanics and make public
