+++ setup.py
@@ -102,7 +102,7 @@ modules = [
     'sympy.parsing.latex._antlr',
     'sympy.parsing.latex.lark',
     'sympy.physics',
-    'sympy.physics._biomechanics',
+    'sympy.physics.biomechanics',
     'sympy.physics.continuum_mechanics',
     'sympy.physics.control',
     'sympy.physics.hep',
@@ -261,7 +261,7 @@ tests = [
     'sympy.multipledispatch.tests',
     'sympy.ntheory.tests',
     'sympy.parsing.tests',
-    'sympy.physics._biomechanics.tests',
+    'sympy.physics.biomechanics.tests',
     'sympy.physics.continuum_mechanics.tests',
     'sympy.physics.control.tests',
     'sympy.physics.hep.tests',
