+++ sympy/core/tests/test_args.py
@@ -3362,48 +3362,48 @@ def test_sympy__physics__vector__frame__CoordinateSym():
 
 
 @SKIP("abstract class")
-def test_sympy__physics___biomechanics__curve__CharacteristicCurveFunction():
+def test_sympy__physics__biomechanics__curve__CharacteristicCurveFunction():
     pass
 
 
-def test_sympy__physics___biomechanics__curve__TendonForceLengthDeGroote2016():
-    from sympy.physics._biomechanics import TendonForceLengthDeGroote2016
+def test_sympy__physics__biomechanics__curve__TendonForceLengthDeGroote2016():
+    from sympy.physics.biomechanics import TendonForceLengthDeGroote2016
     l_T_tilde, c0, c1, c2, c3 = symbols('l_T_tilde, c0, c1, c2, c3')
     assert _test_args(TendonForceLengthDeGroote2016(l_T_tilde, c0, c1, c2, c3))
 
 
-def test_sympy__physics___biomechanics__curve__TendonForceLengthInverseDeGroote2016():
-    from sympy.physics._biomechanics import TendonForceLengthInverseDeGroote2016
+def test_sympy__physics__biomechanics__curve__TendonForceLengthInverseDeGroote2016():
+    from sympy.physics.biomechanics import TendonForceLengthInverseDeGroote2016
     fl_T, c0, c1, c2, c3 = symbols('fl_T, c0, c1, c2, c3')
     assert _test_args(TendonForceLengthInverseDeGroote2016(fl_T, c0, c1, c2, c3))
 
 
-def test_sympy__physics___biomechanics__curve__FiberForceLengthPassiveDeGroote2016():
-    from sympy.physics._biomechanics import FiberForceLengthPassiveDeGroote2016
+def test_sympy__physics__biomechanics__curve__FiberForceLengthPassiveDeGroote2016():
+    from sympy.physics.biomechanics import FiberForceLengthPassiveDeGroote2016
     l_M_tilde, c0, c1 = symbols('l_M_tilde, c0, c1')
     assert _test_args(FiberForceLengthPassiveDeGroote2016(l_M_tilde, c0, c1))
 
 
-def test_sympy__physics___biomechanics__curve__FiberForceLengthPassiveInverseDeGroote2016():
-    from sympy.physics._biomechanics import FiberForceLengthPassiveInverseDeGroote2016
+def test_sympy__physics__biomechanics__curve__FiberForceLengthPassiveInverseDeGroote2016():
+    from sympy.physics.biomechanics import FiberForceLengthPassiveInverseDeGroote2016
     fl_M_pas, c0, c1 = symbols('fl_M_pas, c0, c1')
     assert _test_args(FiberForceLengthPassiveInverseDeGroote2016(fl_M_pas, c0, c1))
 
 
-def test_sympy__physics___biomechanics__curve__FiberForceLengthActiveDeGroote2016():
-    from sympy.physics._biomechanics import FiberForceLengthActiveDeGroote2016
+def test_sympy__physics__biomechanics__curve__FiberForceLengthActiveDeGroote2016():
+    from sympy.physics.biomechanics import FiberForceLengthActiveDeGroote2016
     l_M_tilde, c0, c1, c2, c3, c4, c5, c6, c7, c8, c9, c10, c11 = symbols('l_M_tilde, c0:12')
     assert _test_args(FiberForceLengthActiveDeGroote2016(l_M_tilde, c0, c1, c2, c3, c4, c5, c6, c7, c8, c9, c10, c11))
 
 
-def test_sympy__physics___biomechanics__curve__FiberForceVelocityDeGroote2016():
-    from sympy.physics._biomechanics import FiberForceVelocityDeGroote2016
+def test_sympy__physics__biomechanics__curve__FiberForceVelocityDeGroote2016():
+    from sympy.physics.biomechanics import FiberForceVelocityDeGroote2016
     v_M_tilde, c0, c1, c2, c3 = symbols('v_M_tilde, c0, c1, c2, c3')
     assert _test_args(FiberForceVelocityDeGroote2016(v_M_tilde, c0, c1, c2, c3))
 
 
-def test_sympy__physics___biomechanics__curve__FiberForceVelocityInverseDeGroote2016():
-    from sympy.physics._biomechanics import FiberForceVelocityInverseDeGroote2016
+def test_sympy__physics__biomechanics__curve__FiberForceVelocityInverseDeGroote2016():
+    from sympy.physics.biomechanics import FiberForceVelocityInverseDeGroote2016
     fv_M, c0, c1, c2, c3 = symbols('fv_M, c0, c1, c2, c3')
     assert _test_args(FiberForceVelocityInverseDeGroote2016(fv_M, c0, c1, c2, c3))
