+++ sympy/physics/biomechanics/__init__.py
