+++ sympy/physics/biomechanics/_mixin.py
