+++ sympy/physics/biomechanics/activation.py
@@ -18,7 +18,7 @@ from sympy.core.symbol import Symbol
 from sympy.core.numbers import Float, Integer, Rational
 from sympy.functions.elementary.hyperbolic import tanh
 from sympy.matrices.dense import MutableDenseMatrix as Matrix, zeros
-from sympy.physics._biomechanics._mixin import _NamedMixin
+from sympy.physics.biomechanics._mixin import _NamedMixin
 from sympy.physics.mechanics import dynamicsymbols
